ローカル環境でも HTTP Server を起動できるように main.go を Lambda 用とローカル用に2つ作成
