+++ Makefile
@@ -17,7 +17,7 @@ deps:
 
 .PHONY: build
 build: deps
-	GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -trimpath -o bin/lambda
+	GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -trimpath -o bin/lambda ./cmd/lambda/main.go
 
 .PHONY: zip
 zip: build
