+++ infrastructure/new_uploader.go
@@ -1,0 +1,23 @@
+package infrastructure
+
+import (
+	"context"
+	"log"
+	"os"
+
+	"github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+)
+
+func NewUploader() *manager.Uploader {
+	region := os.Getenv("REGION")
+	ctx := context.Background()
+	cfg, err := config.LoadDefaultConfig(ctx, config.WithRegion(region))
+	if err != nil {
+		// TODO ここでエラーが発生した場合、致命的な問題が起きているのでちゃんとしたログを出すように改修する
+		log.Fatalln(err)
+	}
+	s3Client := s3.NewFromConfig(cfg)
+	return manager.NewUploader(s3Client)
+}
